Pass Weffc++ among other warnings

+++ include/lug/Core/Version.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <cstdint>
+
 namespace lug {
 namespace Core {
 
@@ -12,7 +14,7 @@ struct Version {
     uint32_t patch : 12;
 
     operator uint32_t() const {
-        return (major << 22) | (minor << 12) | patch;
+        return static_cast<uint32_t>(major << 22) | static_cast<uint32_t>(minor << 12) | patch;
     }
 
     static Version fromInt(uint32_t version);
